fix: move work route retirement tests to server_factory_test
Avoid backend-size violation on server_submit_work_test.go (1000 line
limit) while keeping pkg/api at 15 files after deleting the standalone
unscoped routes test file.
